+++ packages/jaspr/lib/src/browser/browser_binding.dart
@@ -10,8 +10,6 @@ import '../framework/framework.dart';
 import 'dom_render_object.dart';
 import 'js_data.dart';
 
-final _queryReg = RegExp(r'^(.*?)(?:\((\d+):(\d+)\))?$');
-
 /// Global component binding for the browser
 class BrowserAppBinding extends AppBinding with ComponentsBinding {
   @override
@@ -21,22 +19,23 @@ class BrowserAppBinding extends AppBinding with ComponentsBinding {
   Uri get currentUri => Uri.parse(window.location.href.substring(window.location.origin.length));
 
   late String attachTarget;
+  late (Node, Node)? attachBetween;
 
   @override
-  Future<void> attachRootComponent(Component app, {String attachTo = 'body'}) {
+  Future<void> attachRootComponent(Component app, {String attachTo = 'body', (Node, Node)? attachBetween}) {
     _loadRawState();
     attachTarget = attachTo;
+    this.attachBetween = attachBetween;
     return super.attachRootComponent(app);
   }
 
   @override
   RenderObject createRootRenderObject() {
-    var attachMatch = _queryReg.firstMatch(attachTarget)!;
-    var target = attachMatch.group(1)!;
-    var from = int.tryParse(attachMatch.group(2) ?? '');
-    var to = int.tryParse(attachMatch.group(3) ?? '');
-
-    return RootDomRenderObject(document.querySelector(target)!, from, to);
+    if (attachBetween case (var start, var end)) {
+      return RootDomRenderObject.between(start, end);
+    } else {
+      return RootDomRenderObject(document.querySelector(attachTarget)!);
+    }
   }
 
   final Map<String, dynamic> _rawState = {};