+++ packages/jaspr/lib/src/server/server_binding.dart
@@ -1,15 +1,14 @@
 import 'dart:async';
 import 'dart:convert';
 
-import '../foundation/basic_types.dart';
-import '../foundation/binding.dart';
-import '../foundation/options.dart';
-import '../framework/framework.dart';
-import 'document/document.dart';
+import '../../jaspr.dart';
+import 'adapters/client_component_adapter.dart';
+import 'adapters/document_adapter.dart';
+import 'adapters/sync_script_adapter.dart';
 import 'markup_render_object.dart';
 
 /// Global component binding for the server
-class ServerAppBinding extends AppBinding with ComponentsBinding, DocumentBinding {
+class ServerAppBinding extends AppBinding with ComponentsBinding {
   @override
   Uri get currentUri => _currentUri!;
   Uri? _currentUri;
@@ -31,8 +30,25 @@ class ServerAppBinding extends AppBinding with ComponentsBinding, DocumentBindin
   Future<String> render() async {
     await rootCompleter.future;
 
-    var renderObject = rootElement!.renderObject as MarkupRenderObject;
-    return renderDocument(renderObject);
+    var root = rootElement!.renderObject as MarkupRenderObject;
+    var adapters = [
+      ..._adapters.reversed,
+      SyncScriptAdapter(getStateData),
+      DocumentAdapter(),
+    ];
+
+    for (var adapter in adapters.reversed) {
+      var r = adapter.prepare();
+      if (r is Future) {
+        await r;
+      }
+    }
+
+    for (var adapter in adapters) {
+      adapter.apply(root);
+    }
+
+    return root.renderToHtml();
   }
 
   Future<String> data() async {
@@ -61,10 +77,35 @@ class ServerAppBinding extends AppBinding with ComponentsBinding, DocumentBindin
     return MarkupRenderObject();
   }
 
+  late Future<String> Function(String) _fileHandler;
+
+  void setFileHandler(Future<String> Function(String) handler) {
+    _fileHandler = handler;
+  }
+
+  Future<String> loadFile(String name) => _fileHandler(name);
+
+  late final List<RenderAdapter> _adapters = [];
+
+  void addRenderAdapter(RenderAdapter adapter) {
+    _adapters.add(adapter);
+  }
+
   JasprOptions get options => _options!;
   JasprOptions? _options;
 
   void initializeOptions(JasprOptions options) {
     _options = options;
   }
+
+  @override
+  Future<void> attachRootComponent(Component app) {
+    return super.attachRootComponent(ClientComponentRegistry(child: app));
+  }
+}
+
+abstract class RenderAdapter {
+  FutureOr<void> prepare();
+
+  void apply(MarkupRenderObject root);
 }