+++ packages/jaspr/CHANGELOG.md
@@ -1,6 +1,8 @@
-## Unreleased patch
+## Unreleased minor
 
 - Fixed bug with `DomValidator`.
+- `Document` is no longer required when using server-side rendering.
+- Improved how `@client` components are hydrated.
 
 ## 0.10.0